reset: trial to cleanup reset handling and use reset_simple
This commit is an ongoing trial to use more heavily the reset_simple
driver in order to perform the reset handling in STi platform.
I am not happy with the current reset driver since it hardcode the
compatible of the syscfg area within the driver.  Having a deeper
look in it, it turns out that this is quite simple handling which
could as well be done by the reset simple driver.

The missing point is the powerdown driver which is needs to poll
for a ack register in order to assert / deassert the reset, which
is something much more complication than what the reset_simple driver
currently does.
Hence for the time being this is just partially done here but not
applied.

Signed-off-by: Alain Volmat <avolmat@me.com>