dma: st: fdma: correct unbalanced rproc boot/shutdown
The FDMA driver is starting the remote proc via the st_slim_rproc
code by relying on rproc_alloc.  Since auto_boot is default to true
in rproc_alloc and the st_slim_rproc doesn't affect this variable,
the rproc firmware for the FDMA is automatically started as soon as
the firmware becomes available.

Add a check against state of the rproc to ensure it is already
running at that time, otherwise DEFER the execution and remove
rproc_shutdown call from the channel free function since otherwise
this would lead to unbalanced call of boot/shutdown.
The rproc is already stop as part of the rproc_del function called
at remove time.

Signed-off-by: Alain Volmat <avolmat@me.com>