drm: sti: add support for VTG on the stih418 platform
VTG integrated into the STiH418 differ in the number of outputs
available and allocation of each output.  Indeed on STiH418, there
are 6 outputs (4 on the STiH407/STiH410) and HDMI is connected to
the 5th output in case of STiH418 while it is on the 1st output
in case of STiH407/STiH410.

Two new compatibles st,stih407-vtg and st,stih418-vtg are added
to differentiate the implementation, while st,vtg is kept for
legacy, behaving as st,stih407-vtg.

Signed-off-by: Alain Volmat <avolmat@me.com>