drm: sti: add compositor support for stih418 platform
On the STiH418, a new clock (proc_mixer) must be enabled in order
to have the plane mixers properly behaving.  Add a new
st,stih418-compositor in order to describe the planes/mixers
available on this platform.  For the time being keep it simple with
only 4 planes and 2 mixers, others will be added in further commits.

Signed-off-by: Alain Volmat <avolmat@me.com>