gpu: drm: sti: add support for GDPPLUS / stih418 GDPs
The STiH418 platform embeds two kinds of graphical planes (GDP),
so called GDPPLUS which has additional (yet unimplemented)
features compared to the GDP, and also the GDP.
Register map of GDPPLUS slightly differ from the GDP even if,
for common functionalities registers name and behavior are the
same (but not at same addresses).  For that purpose, this commit
rework the way of addressing the GDP(PLUS) registers.
For the time being only GDPPLUS are implemented in case of
STiH418, GDPs will be added in a further commit.

Signed-off-by: Alain Volmat <avolmat@me.com>