tty: serial: st-asc: fix race between startup / console_write
The startup ops is protected by the uart_port mutex, while the
console_write is protected by the uart_port spin_lock.  This leads
to a race condition situation when the startup call comes in the
middle of a console_write.  Indeed during the console_write, the
interrupts are disabled and the rx interrupt enabling performed
by the startup op can thus be override when enabling back the
interrupts at the end of console_write.

To fix this issue, protect the asc_enable_rx_interrupts call with the
uart_port spin_lock.

Signed-off-by: Alain Volmat <alain.volmat@foss.st.com>
Change-Id: I6409c5461bf1ebcf2293fb1fe6d4f65d5e2dd6b6
Reviewed-on: https://gerrit.st.com/c/catapult-os/kernel/+/310137
Tested-by: Jerome AUDU <jerome.audu@st.com>
Reviewed-by: Jerome AUDU <jerome.audu@st.com>
Reviewed-by: Raphael GALLAIS-POU <raphael.gallais-pou@foss.st.com>