drm: sti: use ports entries to select match component
In order to avoid trying to match unavailable component
use the ports properties of the display-subsystem node
to select which driver to match for.
That done, it becomes possible to put the display-substem
node out of the soc node within the device-tree and not
have each IP blocks as children of the display-subsystem
node.
The driver will always match several components (tvout
and compositor), while others (dvo, hda, hdmi) depends
on the output selected via the ports properties.

Signed-off-by: Alain Volmat <avolmat@me.com>